Restructuring of tabor driver code
